LPD-22309 Supporting DXP versions as virtual items
